+++ controllers/contacts/index.js
@@ -2,12 +2,12 @@ const getAll = require("./getAll");
 const getById = require("./getById");
 const add = require("./add");
 const updateById = require("./updateById");
-const removeContact = require("./removeById");
+const removeById = require("./removeById");
 
 module.exports = {
   getAll,
   getById,
   add,
   updateById,
-  removeContact,
+  removeById,
 };
