+++ helpers/ctrlWrapper.js
@@ -2,7 +2,9 @@ const ctrlWrapper = (ctrl) => {
   const func = async (req, res, next) => {
     try {
       await ctrl(req, res, next);
-    } catch (error) {}
+    } catch (error) {
+      next(error);
+    }
   };
   return func;
 };
