+++ middlewares/validateBody.js
@@ -1,8 +1,8 @@
 const { RequestError } = require("../helpers");
 
-const validateBody = (contactsSchema) => {
+const validateBody = (schema) => {
   const func = (req, res, next) => {
-    const { error } = contactsSchema.validate(req.body);
+    const { error } = schema.validate(req.body);
     if (error) {
       next(RequestError(400, "missing required name field"));
     }
