+++ schemas/contactsSchema.js
@@ -1,6 +1,6 @@
 const Joi = require("joi");
 
-const contactsSchema = Joi.object({
+const addSchema = Joi.object({
   name: Joi.string().min(3).max(30).required(),
   email: Joi.string()
     .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
@@ -11,4 +11,6 @@ const contactsSchema = Joi.object({
     .required(),
 });
 
-module.exports = contactsSchema;
+module.exports = {
+  addSchema,
+};
